added classes and render item
