+++ core/src/com/mygdx/game/Level.java
@@ -1,42 +1,4 @@
 package com.mygdx.game;
 
-import com.badlogic.gdx.Screen;
-
-public class Level implements Screen {
-    private VMGame game;
-
-    @Override
-    public void show() {
-
-    }
-
-    @Override
-    public void render(float delta) {
-
-    }
-
-    @Override
-    public void resize(int width, int height) {
-
-    }
-
-    @Override
-    public void pause() {
-
-    }
-
-    @Override
-    public void resume() {
-
-    }
-
-    @Override
-    public void hide() {
-
-    }
-
-    @Override
-    public void dispose() {
-
-    }
+public class Level {
 }
